long

+++ Packages_2024/gmp_6.3.0/fp_gmp.pas
@@ -24,7 +24,7 @@ type
 
   {$IFDEF Windows}
   Tculong = uint32;
-  Tclong = int64;
+  Tclong = int32;
   Tlong_double = double;
   {$ENDIF}
   Pculong = ^Tculong;
