+++ Packages_2024/gsl_2.7.1/fp_gsl.pas
@@ -25,7 +25,7 @@ type
   {$ENDIF}
   {$IFDEF windows}
   Tculong = uint32;
-  Tclong = int64;
+  Tclong = int32;
   Tlong_double = double;
   {$ENDIF}
   Pculong = ^Tculong;
