+++ Packages_2024/pcap_1.10.4/fp_pcap.pas
@@ -25,7 +25,7 @@ type
   {$ENDIF}
   {$IFDEF windows}
   Tculong = uint32;
-  Tclong = int64;
+  Tclong = int32;
   {$ENDIF}
 
 type
