+++ Packages_2024/tcl-8.6.14/fp_tcl.pas
@@ -33,7 +33,7 @@ type
   {$ENDIF}
   {$IFDEF windows}
   Tculong = uint32;
-  Tclong = int64;
+  Tclong = int32;
   {$ENDIF}
   Pculong = ^Tculong;
   Pclong = ^Tclong;
