+++ Packages_2024/tommath/fp_tommath_1_2_1.pas
@@ -26,7 +26,7 @@ type
   {$ENDIF}
   {$IFDEF windows}
   Tculong = uint32;
-  Tclong = int64;
+  Tclong = int32;
   {$ENDIF}
   Pculong = ^Tculong;
   Pclong = ^Tclong;
